+++ index.js
@@ -1,13 +1,9 @@
-const bboxPolygon = require('@turf/bbox-polygon')
-const explode = require('@turf/explode')
-const inside = require('@turf/inside')
+// const bboxPolygon = require('@turf/bbox-polygon')
+// const explode = require('@turf/explode')
+// const inside = require('@turf/inside')
 const turfBBox = require('@turf/bbox')
-const normalize = require('geojson-normalize')
-const mercator = require('global-mercator')
-const hash = mercator.hash
-const range = mercator.range
-const tileToBBox = mercator.tileToBBox
-const lngLatToTile = mercator.lngLatToTile
+const {hash, range, lngLatToTile} = require('global-mercator')
+const {featureEach} = require('@turf/meta')
 
 /**
  * Creates an Iterator of Tiles from a given BBox
@@ -27,47 +23,39 @@ function * single (extent, minZoom, maxZoom) {
   for (const [columns, rows, zoom] of levels(extent, minZoom, maxZoom)) {
     for (const row of rows) {
       for (const column of columns) {
-        // Filter by Unique key
         const tile = [column, row, zoom]
         const key = hash(tile)
 
+        // Only return unique key
         if (!unique[key]) {
           unique[key] = true
-
-          // Filter geospatially
-          if (extent.type === 'Feature' || extent.type === 'FeatureCollection') {
-            let isInside = false
-            const geojson = normalize(extent)
-            const bbox = tileToBBox(tile)
-            const polygon = bboxPolygon(bbox)
-            const exploded = explode(polygon)
-
-            // Remove any GeoJSON that do not meet zoom level requirements
-            geojson.features = geojson.features.filter(feature => {
-              const featureMinZoom = feature.properties.minZoom || feature.properties.minzoom
-              const featureMaxZoom = feature.properties.maxZoom || feature.properties.maxzoom
-              if (zoom < featureMinZoom) { return false }
-              if (zoom > featureMaxZoom) { return false }
-              return true
-            })
-
-            for (const feature of geojson.features) {
-              if (isInside) { break }
-              for (const point of exploded.features) {
-                if (inside(point, feature)) {
-                  isInside = true
-                  break
-                }
-              }
-            }
-            // Return tile if inside GeoJSON
-            if (isInside) { yield tile }
-          // Return tile if not GeoJSON
-          } else { yield tile }
+          yield tile
         }
       }
     }
   }
+}
+
+/**
+ * All Tiles from a given BBox
+ *
+ * @param {BBox|BBox[]|GeoJSON} extent BBox [west, south, east, north] order or GeoJSON Polygon
+ * @param {number} minZoom Minimum Zoom
+ * @param {number} maxZoom Maximum Zoom
+ * @returns {Array<Tile>} Tiles from extent
+ * @example
+ * const tiles = slippyGrid.all([-180.0, -90.0, 180, 90], 3, 8)
+ * //=tiles
+ */
+function all (extent, minZoom, maxZoom) {
+  const tiles = []
+  const grid = single(extent, minZoom, maxZoom)
+  while (true) {
+    const {value, done} = grid.next()
+    if (done) break
+    tiles.push(value)
+  }
+  return tiles
 }
 
 /**
@@ -118,6 +106,9 @@ function * bulk (extent, minZoom, maxZoom, size) {
  */
 function levels (extent, minZoom, maxZoom) {
   const extents = []
+  if (extent === undefined) throw new Error('extent is required')
+  if (minZoom === undefined) throw new Error('minZoom is required')
+  if (maxZoom === undefined) throw new Error('maxZoom is required')
 
   // Single Array
   if (extent.length === 4 && extent[0][0] === undefined) { extents.push({bbox: extent, minZoom, maxZoom}) }
@@ -126,15 +117,12 @@ function levels (extent, minZoom, maxZoom) {
   if (extent.length && extent[0][0] !== undefined) { extent.map(inner => extents.push({bbox: inner, minZoom, maxZoom})) }
 
   // GeoJSON
-  if (extent.type === 'Feature' || extent.type === 'FeatureCollection') {
-    const geojson = normalize(extent)
-    geojson.features.map(feature => {
-      const bbox = turfBBox(feature)
-      const featureMinZoom = feature.properties.minZoom || feature.properties.minzoom || minZoom
-      const featureMaxZoom = feature.properties.maxZoom || feature.properties.maxzoom || maxZoom
-      extents.push({bbox, minZoom: featureMinZoom, maxZoom: featureMaxZoom})
-    })
-  }
+  featureEach(extent, feature => {
+    const bbox = turfBBox(feature)
+    const featureMinZoom = feature.properties.minZoom || feature.properties.minzoom || minZoom
+    const featureMaxZoom = feature.properties.maxZoom || feature.properties.maxzoom || maxZoom
+    extents.push({bbox, minZoom: featureMinZoom, maxZoom: featureMaxZoom})
+  })
   const levels = []
   for (const {bbox, minZoom, maxZoom} of extents) {
     let [x1, y1, x2, y2] = bbox
@@ -143,12 +131,32 @@ function levels (extent, minZoom, maxZoom) {
       if (y2 < -85) y2 = -85
       const t1 = lngLatToTile([x1, y1], zoom)
       const t2 = lngLatToTile([x2, y2], zoom)
+
+      // Columns
+      let columns = []
+      // Fiji - World divided into two parts
+      if (t1[0] > t2[0]) {
+        // right world +180 degrees
+        const maxtx = Math.pow(2, zoom) - 1
+        const rightColumns = range(t1[0], maxtx + 1)
+        rightColumns.forEach(column => columns.push(column))
+
+        // left world -180 degrees
+        const mintx = 0
+        const leftColumns = range(mintx, t2[0] + 1)
+        leftColumns.forEach(column => columns.push(column))
+      } else {
+        // Normal World
+        const mintx = Math.min(t1[0], t2[0])
+        const maxtx = Math.max(t1[0], t2[0])
+        columns = range(mintx, maxtx + 1)
+      }
+
+      // Rows
       const minty = Math.min(t1[1], t2[1])
       const maxty = Math.max(t1[1], t2[1])
-      const mintx = Math.min(t1[0], t2[0])
-      const maxtx = Math.max(t1[0], t2[0])
       const rows = range(minty, maxty + 1)
-      const columns = range(mintx, maxtx + 1)
+
       levels.push([columns, rows, zoom])
     }
   }
@@ -172,10 +180,12 @@ function count (extent, minZoom, maxZoom, quick) {
   let count = 0
 
   // Quick count
-  for (const [columns, rows] of levels(extent, minZoom, maxZoom)) {
-    count += rows.length * columns.length
+  if (quick !== -1) {
+    for (const [columns, rows] of levels(extent, minZoom, maxZoom)) {
+      count += rows.length * columns.length
+    }
+    if (count > quick) { return count }
   }
-  if (count > quick) { return count }
 
   // Accurate count
   count = 0
@@ -187,4 +197,4 @@ function count (extent, minZoom, maxZoom, quick) {
   }
   return count
 }
-module.exports = {single, bulk, levels, count}
+module.exports = {single, bulk, levels, count, all}