+++ test.js
@@ -4,7 +4,7 @@ const load = require('load-json-file')
 const mercator = require('global-mercator')
 const write = require('write-json-file')
 const turf = require('@turf/turf')
-const normalize = require('geojson-normalize')
+const featureEach = require('@turf/meta').featureEach
 const slippyGrid = require('.')
 
 const directories = {
@@ -21,24 +21,23 @@ for (const filename of fs.readdirSync(directories.in)) {
   fixtures.push(fixture)
 }
 
+const FIJI = load.sync(path.join(__dirname, 'test', 'in', 'fiji.geojson'))
 const WORLD = [-180.0, -90.0, 180, 90]
 
 /**
  * Grid
  */
-describe('grid', () => {
+describe('slippy-grid', () => {
   test('single', () => {
     const grid = slippyGrid.single(WORLD, 3, 21)
     expect(grid.next().value).toEqual([0, 0, 3])
   })
 
-  for (const fixture of fixtures) {
-    const name = fixture.name
-    const geojson = normalize(JSON.parse(JSON.stringify(fixture.geojson)))
-    const collection = turf.featureCollection([])
+  for (const {geojson, name} of fixtures) {
+    const results = turf.featureCollection([])
 
     test('single ' + name, () => {
-      const grid = slippyGrid.single(JSON.parse(JSON.stringify(fixture.geojson)), 1, 8)
+      const grid = slippyGrid.single(geojson, 4, 6)
       while (true) {
         const {value, done} = grid.next()
         if (done) { break }
@@ -52,16 +51,16 @@ describe('grid', () => {
           z: tile[2],
           bbox
         }
-        collection.features.push(polygon)
+        results.features.push(polygon)
       }
-      geojson.features.map(feature => {
+      featureEach(geojson, feature => {
         feature.properties['fill-opacity'] = 0.2
         feature.properties['stroke'] = '#0000ff'
         feature.properties['stroke-width'] = 3
-        collection.features.push(feature)
+        results.features.push(feature)
       })
-      if (process.env.REGEN) { write.sync(directories.out + name, collection) }
-      expect(collection, load.sync(directories.out + name))
+      if (process.env.REGEN) { write.sync(directories.out + name, results) }
+      expect(results, load.sync(directories.out + name))
     })
   }
 
@@ -91,7 +90,11 @@ describe('grid', () => {
     expect(value.length).toBe(32)
   })
 
-  test('count', () => expect(slippyGrid.count(WORLD, 3, 21)).toBe(5854460328256))
+  test('count', () => expect(slippyGrid.count(WORLD, 3, 21)).toBe(5844858641728))
   test('levels', () => expect(slippyGrid.levels(WORLD, 3, 21).length).toBe(19))
   test('levels.world.zoom1', () => expect(slippyGrid.levels(WORLD, 1, 1)).toEqual([[[0, 1], [0, 1], 1]]))
+  test('fiji', () => {
+    const tiles = slippyGrid.all(FIJI, 2, 2)
+    expect(tiles).toEqual([[3, 1, 2], [0, 1, 2]])
+  })
 })